新関数に切り替え。引数が変わった
